enhance(frontend): organize settings page
